requirements.txt
langchain==0.3.14
langchain-anthropic==0.3.3
langchain-aws==0.2.11
langchain-chroma==0.2.0
langchain-community==0.3.14
langchain-core==0.3.31
langchain-fireworks==0.2.6
langchain-google-community==2.0.4
langchain-google-genai==2.0.8
langchain-ollama==0.2.2
langchain-openai==0.3.1
langchain-text-splitters==0.3.5
langchain-unstructured==0.1.6
langchainhub==0.1.15
openai>=1.12.0
faiss-cpu>=1.7.4
pydantic>=2.0.0
typing-extensions>=4.5.0
python-dotenv>=1.0.0
nest-asyncio>=1.5.8
beautifulsoup4>=4.12.0
playwright>=1.41.0
duckduckgo-search>=4.1.1
pyowm>=3.3.0